+++ R/package.R
@@ -8,6 +8,12 @@
 #' algorithm is also provided.
 #'
 #' @details
+#' Both the MP and OMP algorithms only support Gabor atoms. However, both algorithms
+#' are much more general. They can handle any atom dictionary, as long as we can
+#' compute the dot products of the signal and the atoms. Gabor atoms are particularly
+#' popular because they well implement the time-frequency tradeoff implied by the
+#' Heisenberg Uncertainty Principle and describe many natural signals.
+#'
 #' In addition to generic time_series data, the package supports direct loading of
 #' data stored in \strong{EDF/EDF(+)} and \strong{WFDB} (WaveForm DataBase) formats.
 #' These formats are widely used for physiological signals such as EEG and ECG recordings.
@@ -19,18 +25,18 @@
 #' \strong{Piotr T. Różański} and is available at
 #' \url{https://github.com/develancer/empi}
 #'
-#' Example datasets available via the \code{system_file()} function:
+#' Example datasets available via the \code{system.file()} function:
 #'
 #'  \itemize{
-#'    \item \code{EEG.edf}, \code{EEG.db}, \code{EEG.csv}
+#'    \item \code{EEG.edf}
 #'      \itemize{
 #'        \item 19 EEG channels + 1 EDF_Annotations channel
 #'        \item sampling frequency: 256 Hz, signal length: 10 sec.
 #'        \item channel names: Fp1, Fp2, F3, F4, F7, F8, Fz, C3, C4, Cz, T3, T5, T4, T6, P3, P4, Pz, O1, O2, EDF_Annotations
 #'      }
-#'    \item \code{EEG_bipolar_filtered.db}
+#'    \item \code{EEG_bipolar_filtered.db}, \code{EEG_bipolar_filtered.csv}, \code{EEG_bipolar_filtered.bin}
 #'      \itemize{
-#'        \item 18 EEG channels after double_banana montage and filtering of the data from \code{EEG.edf}
+#'        \item 18 EEG channels after application of the double-banana montage and filtering of the \code{EEG.edf} data
 #'        \item sampling frequency: 256 Hz, signal length: 10 sec.
 #'        \item channel names: Fp2_F4, F4_C4, C4_P4, P4_O2, Fp1_F3, F3_C3, C3_P3, P3_O1, Fp2_F8, F8_T4,
 #'                             T4_T6, T6_O2, Fp1_F7, F7_T3, T3_T5, T5_O1, Fz_Cz, Cz_Pz
@@ -56,7 +62,8 @@
 #'        \item 12 ECG leads, 10 sec, 16-bit integer format
 #'        \item standard lead names: I, II, III, aVR, aVL, aVF, V1–V6
 #'      }
-#'    \item \code{sample1_dict.xml}, \code{sample2_dict.xml}, \code{sample3_dict.xml}, \code{EEG_dict.xml}
+#'    \item \code{sample1_dict.xml}, \code{sample2_dict.xml}, \code{sample3_dict.xml}, \code{sample3_dict_EMPI.xml},
+#'    \code{EEG_bipolar_filtered.xml}, \code{sample1_dict_one_block.xml}
 #'      \itemize{
 #'        \item XML files describing a multiscale Gabor dictionary.
 #'        \item such files can be generated from the EMPI program executed with the \code{--dictionary-output}