+++ R/read_empi_db_file.R
@@ -20,7 +20,7 @@
 #'
 #' @examples
 #' ## Not run:
-#' file <- system.file("extdata", "EEG.db", package = "MatchingPursuit")
+#' file <- system.file("extdata", "EEG_bipolar_filtered.db", package = "MatchingPursuit")
 #' out <- read_empi_db_file(file)
 #'
 #' n_channels <- ncol(out$original_signal)